converted sql files to universal SQL

+++ tkp/database/sql/statements/tables/assoccatsource.sql
@@ -1,10 +1,10 @@
 CREATE TABLE assoccatsource
   (xtrsrc INT NOT NULL
   ,catsrc INT NOT NULL
-  ,type TINYINT NOT NULL
-  ,distance_arcsec DOUBLE NULL
-  ,r DOUBLE NULL
-  ,loglr DOUBLE NULL
+  ,type SMALLINT NOT NULL
+  ,distance_arcsec DOUBLE PRECISION NULL
+  ,r DOUBLE PRECISION NULL
+  ,loglr DOUBLE PRECISION NULL
   ,PRIMARY KEY (xtrsrc ,catsrc)
   ,FOREIGN KEY (xtrsrc) REFERENCES extractedsource (id)
   ,FOREIGN KEY (catsrc) REFERENCES catalogedsource (id)
