+++ tkp/database/sql/statements/tables/assocskyrgn.sql
@@ -5,7 +5,7 @@
 CREATE TABLE assocskyrgn
   (runcat INT NOT NULL
   ,skyrgn INT NOT NULL
-  ,distance_deg DOUBLE 
+  ,distance_deg DOUBLE PRECISION
   ,FOREIGN KEY (runcat) REFERENCES runningcatalog (id)
   ,FOREIGN KEY (skyrgn) REFERENCES skyregion (id)
 );
