+++ tkp/database/sql/statements/tables/assocxtrsource.sql
@@ -1,10 +1,10 @@
 CREATE TABLE assocxtrsource
   (runcat INT NOT NULL
   ,xtrsrc INT NULL
-  ,type TINYINT NOT NULL
-  ,distance_arcsec DOUBLE NULL
-  ,r DOUBLE NULL
-  ,loglr DOUBLE NULL
+  ,type SMALLINT NOT NULL
+  ,distance_arcsec DOUBLE PRECISION NULL
+  ,r DOUBLE PRECISION NULL
+  ,loglr DOUBLE PRECISION NULL
   ,PRIMARY KEY (runcat, xtrsrc)
   ,FOREIGN KEY (runcat) REFERENCES runningcatalog (id)
   ,FOREIGN KEY (xtrsrc) REFERENCES extractedsource (id)
