+++ tkp/database/sql/statements/tables/classification.sql
@@ -5,5 +5,5 @@
 CREATE TABLE classification
 	( transient_id INTEGER NOT NULL
 	,classification VARCHAR(256)
-	,weight DOUBLE NOT NULL DEFAULT 0
+	,weight DOUBLE PRECISION NOT NULL DEFAULT 0
 );
