+++ tkp/database/sql/statements/tables/runningcatalog_flux.sql
@@ -1,19 +1,19 @@
 CREATE TABLE runningcatalog_flux
   (runcat INT NOT NULL
   ,band SMALLINT NOT NULL
-  ,stokes TINYINT NOT NULL DEFAULT 1
+  ,stokes SMALLINT NOT NULL DEFAULT 1
   ,f_datapoints INT NOT NULL
-  ,resolution DOUBLE NULL
-  ,avg_f_peak DOUBLE NULL
-  ,avg_f_peak_sq DOUBLE NULL
-  ,avg_f_peak_weight DOUBLE NULL
-  ,avg_weighted_f_peak DOUBLE NULL
-  ,avg_weighted_f_peak_sq DOUBLE NULL
-  ,avg_f_int DOUBLE NULL
-  ,avg_f_int_sq DOUBLE NULL
-  ,avg_f_int_weight DOUBLE NULL
-  ,avg_weighted_f_int DOUBLE NULL
-  ,avg_weighted_f_int_sq DOUBLE NULL
+  ,resolution DOUBLE PRECISION NULL
+  ,avg_f_peak DOUBLE PRECISION NULL
+  ,avg_f_peak_sq DOUBLE PRECISION NULL
+  ,avg_f_peak_weight DOUBLE PRECISION NULL
+  ,avg_weighted_f_peak DOUBLE PRECISION NULL
+  ,avg_weighted_f_peak_sq DOUBLE PRECISION NULL
+  ,avg_f_int DOUBLE PRECISION NULL
+  ,avg_f_int_sq DOUBLE PRECISION NULL
+  ,avg_f_int_weight DOUBLE PRECISION NULL
+  ,avg_weighted_f_int DOUBLE PRECISION NULL
+  ,avg_weighted_f_int_sq DOUBLE PRECISION NULL
   ,PRIMARY KEY (runcat
                ,band
                ,stokes
